Move API-related fixtures to tests/api/conftest.py
